Fix incompatibility with irmago in bigint encodings of attributes
Go's *big.Int.SetBytes() method interprets its argument as an unsigned int, i.e. without interpreting the most signicifant bit as sign bit. By contrast, in Java this bit determines the sign of the resulting bigint. Consequentially, some strings convert to different bigints in Java and Go: for example, in UTF8 the string "é" converts to the bigints -15447 in Java and 50089 in Go. During the string->bigint conversion that happens on both sides during issuance, this can cause issuance of invalid CL-signatures over the attributes. This commit fixes this, ensuring that the input bytes are always interpreted as unsigned and thus positive, as in Go.